+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 #### ERCOT
 * ERCOT Available Resource Planned Outage Capacity 7 Day and Future datasets in [#884](https://github.com/gridstatus/gridstatus/pull/884)
+* ERCOT SCED AS Price Corrections via `Ercot.get_mcpc_sced_price_corrections`
 
 #### MISO
 * MISO Area Control Error dataset via `MISO.get_area_control_error`